JsonLD - Adjusted Thing
